Update nx_joystick.c
The reason for this change is add ZL, ZR and Left Thumb Stick (Aka L3) buttons.  This change is to reflect Mortal Kombat series or other fighter games that uses that standard layout button mapping.  The re-order of the buttons have also been made to reflect the same as used in the popular consoles (XBOX or Playstation).